src/webhook_delivery_worker/requirements.txt
certifi==2025.11.12
charset-normalizer==3.4.4
dnspython==2.8.0
idna==3.11
pika==1.3.2
pymongo==4.15.3
python-dotenv==1.2.1
requests==2.32.5
urllib3==2.5.0
